extract function onlyOfTypes
